# Conflicts:
#	BackEnd/ad-service/src/main/resources/application.yml
#	BackEnd/authentication-service/src/main/java/microservice/authenticationservice/controller/AuthenticationController.java
#	BackEnd/authentication-service/src/main/java/microservice/authenticationservice/service/AuthenticationServiceImpl.java
#	BackEnd/authentication-service/src/main/resources/application.yml
#	BackEnd/bike-service/src/main/resources/application.yml
#	BackEnd/shop-service/src/main/resources/application.yml
#	BackEnd/user-service/src/main/java/microservice/userservice/service/PrivateServiceImpl.java
#	BackEnd/user-service/src/main/resources/application.yml
#	compose.yml

+++ BackEnd/shop-service/src/main/java/microservice/shopservice/model/Shop.java
@@ -36,9 +36,9 @@ public class Shop {
     private String img;
 
     @Column(name = "idUser")
-    private int idUser;
+    private long idUser;
 
-    public Shop( String name, String city, String address, String phoneNumber, String img, int idUser) {
+    public Shop( String name, String city, String address, String phoneNumber, String img, long idUser) {
         this.name = name;
         this.city = city;
         this.address = address;
